chore: moved a test
