+++ tests/test_pocket.py
@@ -2,9 +2,11 @@
 Tests for the Pocket class.
 """
 
+import os
+
 import numpy as np
 
-from deeporigin.drug_discovery import Pocket, Protein
+from deeporigin.drug_discovery import BRD_DATA_DIR, Pocket, Protein
 
 
 def test_pocket_from_ligand():
@@ -43,3 +45,18 @@ def test_pocket_update_coordinates():
     assert np.array_equal(pocket.coordinates, new_coords), (
         "Coordinates should be updated"
     )
+
+
+def test_from_residue_num():
+    """Test creating a pocket from a residue number"""
+
+    # Read in a protein
+    pdb_path = os.path.join(BRD_DATA_DIR, "brd.pdb")
+    protein = Protein.from_file(pdb_path)
+
+    # Create custom pocket
+    custom_pocket = Pocket.from_residue_number(protein, residue_number=77, cutoff=5)
+
+    assert isinstance(
+        custom_pocket.get_center(), np.ndarray
+    ) and custom_pocket.get_center().shape == (3,)
